fix: update precision setting in AccelerationConfig for ProSIP model
- Changed the precision setting in the AccelerationConfig from 'bf16-true' to 'bf16-mixed' to improve compatibility and performance during model training.
- This adjustment aligns with best practices for mixed precision training, enhancing the efficiency of the ProSIP model.

+++ train_prosip.py
@@ -84,7 +84,7 @@ def new_train(
     # Extract acceleration config and create AccelerationConfig instance
     acceleration_config = AccelerationConfig(
         device='auto',
-        precision='bf16-true',
+        precision='bf16-mixed',
         matmul_precision='high',
         compile_model=compile_model
     )
